Added footer component
I added the footer component to the project to give it a full rounded look to the website. Added more skill icons to the skills-page-data.json file. I marked all relevant skills that I use the most with a star icon. I also added a catch-all path for when the user enters a wrong path on my website, it will reroute back to the home page. I added a favicon icon to the website instead of using the generic Angular favicon.