Created ProjectsComponent page
I created the projects component page to showcase all of my skills and accomplishments of the projects. I refactored skills, about me, recommendations, and the home page components to fit the projects data. I renamed the data model files from interface to model to know that they are data models and not normal interfaces. I added the projects-page-data.json file to mimic as if it was a NoSQL database. I moved all the testing files into the respective testing folder to clean up the clutter and make it more organized in the folder structure. I created the services to retrieve specific methods for each page.